backend/apps/core/deps.py
# backend/apps/core/deps.py
# Dependency for getting the current user from the JWT token    
#Adding Middleware for Auth

from fastapi import Depends, HTTPException, status
from fastapi.security import OAuth2PasswordBearer
from jose import JWTError, jwt
from datetime import datetime, timedelta
from apps.core.userstore import fake_user
#from apps.core.settings import SECRET_KEY, ALGORITHM
from apps.core.settings import settings

SECRET_KEY = settings.JWT_SECRET
ALGORITHM = settings.JWT_ALGORITHM

oauth2_scheme = OAuth2PasswordBearer(tokenUrl="auth/login")

def get_current_user(token: str = Depends(oauth2_scheme)) -> str:
    credentials_exception = HTTPException(
        status_code=status.HTTP_401_UNAUTHORIZED,
        detail="Could not validate credentials",
        headers={"WWW-Authenticate": "Bearer"},
    )
    try:
        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
        username: str = payload.get("sub")
        if username is None or username != fake_user["username"]:
            raise credentials_exception
        return username
    except JWTError:
        raise credentials_exception