backend/requirements.txt
# backend/requirements.txt
# Requirements for SamsuBot backend development
asyncpg==0.29.0
passlib[bcrypt]==1.7.4
bcrypt>=3.2.0,<4.1
pydantic-settings

# Web framework & APIcls
fastapi==0.110.0
uvicorn[standard]==0.29.0
httpx>=0.25.2,<1.0.0

# Core language model frameworks
langchain==0.3.27
langchain-community==0.3.21
langsmith>=0.3.45,<0.4.0
#openai==1.30.1

# Vector DB
chromadb==1.0.15

# Tokenizer
tiktoken==0.6.0

# RAG & Embeddings
unstructured==0.14.3
docx2txt==0.8
pdfminer.six==20221105
pypdf==3.17.4
sentence-transformers==5.0.0

# Tools
python-dotenv==1.0.1
pandas==2.2.2
requests==2.31.0
watchfiles==1.1.0
websockets==12.0
pymongo==4.7.0
python-jose[cryptography]==3.3.0
redis==5.0.4
sqlalchemy[asyncio]==2.0.30

# LLM hosting
ollama
langchain-ollama
mistralai # Mistral models      

# Add these for the latest LangChain ecosystem
langchain-chroma
langchain-huggingface